requirements.txt
python-dotenv==1.0.0
streamlit==1.32.0
google-generativeai==0.3.1
pathlib==1.0.1
watchdog==3.0.0